refactor: replace all hooks in hooks folder
