+++ src/lib/hooks/storeHooks.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector, useStore } from 'react-redux';
 import type { TypedUseSelectorHook } from 'react-redux';
-import type { RootState, AppDispatch, AppStore } from './store';
+import type { RootState, AppDispatch, AppStore } from '../store';
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch: () => AppDispatch = useDispatch;
